Renames header file
